// server/src/gql-types.ts
export type IssuesQueryVariables = {
  owner: string,
  name: string,
  cursor?: string | null,
};

export type IssuesQuery = {
  // Lookup a given repository by the owner and repository name.
  repository:  {
    __typename: "Repository",
    // A list of issues that have been opened in the repository.
    issues:  {
      __typename: "IssueConnection",
      // Information to aid in pagination.
      pageInfo:  {
        __typename: "PageInfo",
        // When paginating forwards, the cursor to continue.
        endCursor: string | null,
        // When paginating forwards, are there more items?
        hasNextPage: boolean,
      },
      // A list of nodes.
      nodes:  Array< {
        __typename: "Issue",
        // Identifies the date and time when the object was created.
        createdAt: string,
        // The HTTP URL for this issue
        url: string,
        // `true` if the object is closed (definition of closed may depend on type)
        closed: boolean,
        // A list of events, comments, commits, etc. associated with the issue.
        timeline:  {
          __typename: "IssueTimelineConnection",
          // A list of nodes.
          nodes:  Array<( {
              __typename: "Commit",
            } | {
              __typename: "IssueComment",
            } | {
              __typename: "CrossReferencedEvent",
            } | {
              __typename: "ClosedEvent",
              // Identifies the date and time when the object was created.
              createdAt: string,
            } | {
              __typename: "ReopenedEvent",
            } | {
              __typename: "SubscribedEvent",
            } | {
              __typename: "UnsubscribedEvent",
            } | {
              __typename: "ReferencedEvent",
            } | {
              __typename: "AssignedEvent",
            } | {
              __typename: "UnassignedEvent",
            } | {
              __typename: "LabeledEvent",
            } | {
              __typename: "UnlabeledEvent",
            } | {
              __typename: "MilestonedEvent",
            } | {
              __typename: "DemilestonedEvent",
            } | {
              __typename: "RenamedTitleEvent",
            } | {
              __typename: "LockedEvent",
            } | {
              __typename: "UnlockedEvent",
            }
          ) | null > | null,
        },
      } | null > | null,
    },
  } | null,
};